+++ src/scriptlets/scriptlets-list.js
@@ -67,3 +67,4 @@ export * from './call-nothrow';
 export * from './trusted-create-element';
 export * from './href-sanitizer';
 export * from './json-prune-fetch-response';
+export * from './no-protected-audience';